requirements.txt
opencv-python
streamlit
numpy
Pillow